Restructure the way the modules are made up
