+++ signify/exceptions.py
@@ -1,0 +1,35 @@
+
+
+class SignifyError(Exception):
+    """Base class for all errors generated by Signify"""
+    pass
+
+
+class ParseError(SignifyError):
+    pass
+
+
+class SignerInfoParseError(ParseError):
+    pass
+
+
+class AuthenticodeParseError(ParseError):
+    """Raised when any exception regarding parsing Authenticode structures occurs."""
+    pass
+
+
+class VerificationError(SignifyError):
+    pass
+
+
+class CertificateVerificationError(VerificationError):
+    pass
+
+
+class SignerInfoVerificationError(VerificationError):
+    pass
+
+
+class AuthenticodeVerificationError(VerificationError):
+    """Raised when any exception regarding verifying Authenticode structures occurs."""
+    pass
